requirements.txt
streamlit
pyserial
pandas
openpyxl